+++ packages/orq-rc/RELEASES.md
@@ -9069,3 +9069,13 @@ Based on:
 - [typescript v4.9.0-rc.45] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.45] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.45 - packages/orq-rc
+
+## 2026-05-12 06:54:24
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.46] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.46] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.46 - packages/orq-rc